# docs/api/ezaid-openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogListing": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "entries": {
            "additionalProperties": true,
            "title": "Entries",
            "type": "object"
          },
          "sources": {
            "title": "Sources",
            "type": "string"
          }
        },
        "required": [
          "count",
          "sources",
          "entries"
        ],
        "title": "CatalogListing",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fix": {
            "default": "",
            "title": "Fix",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "GenerationHistory": {
        "properties": {
          "generations": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Generations",
            "type": "array"
          }
        },
        "required": [
          "generations"
        ],
        "title": "GenerationHistory",
        "type": "object"
      },
...
      "Recommendations": {
        "properties": {
          "accelerator": {
            "title": "Accelerator",
            "type": "string"
          },
          "candidates": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Candidates",
            "type": "array"
          },
          "class": {
            "title": "Class",
            "type": "string"
          },
          "group": {
            "title": "Group",
            "type": "string"
          }
        },
        "required": [
          "group",
          "class",
          "accelerator",
          "candidates"
        ],
        "title": "Recommendations",
        "type": "object"
      },
...
      "RollbackOutcome": {
        "properties": {
          "generation": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "rolled_back": {
            "title": "Rolled Back",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "generation",
          "message",
          "rolled_back"
        ],
        "title": "RollbackOutcome",
        "type": "object"
      },
      "RollbackRequest": {
        "properties": {
          "reason": {
            "default": "",
            "title": "Reason",
            "type": "string"
          },
          "reload": {
            "default": false,
            "title": "Reload",
            "type": "boolean"
          },
          "to_generation": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Generation"
          }
        },
        "title": "RollbackRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "serviceToken": {
        "description": "the platform service token (EZAI_CONTROL_TOKEN)",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "One authenticated control plane over the platform's declarative state \u2014 Registry v2 + generations, runtime descriptors, governance queue, rendered artifacts \u2014 for the `local-ezai` CLI (connected mode), the Admin Center and the SWE tool server. Every `/v1` call presents the service token as a bearer credential and may forward the human it acts for (`X-EZAI-User`) and its own name (`X-EZAI-Client`); the audit log records `<user> via <client>`. Mutating calls accept an `Idempotency-Key` header: a retry with the same key and payload replays the stored response, a different payload is a conflict. Every error is `{\"error\": {\"code\", \"message\", \"fix\"}}` \u2014 the same object the CLI prints.",
    "title": "Local-EZAI Platform Control Plane (ezaid)",
    "version": "1.0.0-draft.9",
    "x-contract-status": "draft \u2014 frozen by the P2 phase-close PR"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/catalog": {
      "get": {
        "operationId": "catalog_list",
        "parameters": [
          {
            "description": "forwarded identity of the human the caller acts for",
            "in": "header",
            "name": "X-EZAI-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "forwarded identity of the human the caller acts for",
              "title": "X-Ezai-User"
            }
          },
          {
            "description": "name of the calling surface (cli, admin-center, \u2026)",
            "in": "header",
            "name": "X-EZAI-Client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "name of the calling surface (cli, admin-center, \u2026)",
              "title": "X-Ezai-Client"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogListing"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "missing or invalid service token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "no such model / request / role / project"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "refused by a lifecycle or governance rule (message names the fix)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "invalid request or unservable proposal"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not attached to a platform"
          }
        },
        "security": [
          {
            "serviceToken": []
          }
        ],
        "summary": "Catalog entries (CLI: local-ezai model catalog)",
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/catalog/recommendations": {
      "get": {
        "operationId": "catalog_recommend",
        "parameters": [
          {
            "in": "query",
            "name": "group",
            "required": true,
            "schema": {
              "title": "Group",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "runtime",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Runtime"
            }
          },
          {
            "description": "forwarded identity of the human the caller acts for",
            "in": "header",
            "name": "X-EZAI-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "forwarded identity of the human the caller acts for",
              "title": "X-Ezai-User"
            }
          },
          {
            "description": "name of the calling surface (cli, admin-center, \u2026)",
            "in": "header",
            "name": "X-EZAI-Client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "name of the calling surface (cli, admin-center, \u2026)",
              "title": "X-Ezai-Client"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recommendations"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "missing or invalid service token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "no such model / request / role / project"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "refused by a lifecycle or governance rule (message names the fix)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "invalid request or unservable proposal"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not attached to a platform"
          }
        },
        "security": [
          {
            "serviceToken": []
          }
        ],
        "summary": "Ranked recommendations for a group with fit verdicts (CLI: local-ezai model catalog --group)",
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/generations": {
      "get": {
        "operationId": "generation_history",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "forwarded identity of the human the caller acts for",
            "in": "header",
            "name": "X-EZAI-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "forwarded identity of the human the caller acts for",
              "title": "X-Ezai-User"
            }
          },
          {
            "description": "name of the calling surface (cli, admin-center, \u2026)",
            "in": "header",
            "name": "X-EZAI-Client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "name of the calling surface (cli, admin-center, \u2026)",
              "title": "X-Ezai-Client"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerationHistory"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "missing or invalid service token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "no such model / request / role / project"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "refused by a lifecycle or governance rule (message names the fix)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "invalid request or unservable proposal"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not attached to a platform"
          }
        },
        "security": [
          {
            "serviceToken": []
          }
        ],
        "summary": "Generations with notes and diffs (CLI: local-ezai model history)",
        "tags": [
          "models"
        ]
      }
    },
    "/v1/generations/rollback": {
      "post": {
        "operationId": "generation_rollback",
        "parameters": [
          {
            "description": "forwarded identity of the human the caller acts for",
            "in": "header",
            "name": "X-EZAI-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "forwarded identity of the human the caller acts for",
              "title": "X-Ezai-User"
            }
          },
          {
            "description": "name of the calling surface (cli, admin-center, \u2026)",
            "in": "header",
            "name": "X-EZAI-Client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "name of the calling surface (cli, admin-center, \u2026)",
              "title": "X-Ezai-Client"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RollbackRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollbackOutcome"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "missing or invalid service token"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "no such model / request / role / project"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "refused by a lifecycle or governance rule (message names the fix)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "invalid request or unservable proposal"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not attached to a platform"
          }
        },
        "security": [
          {
            "serviceToken": []
          }
        ],
        "summary": "Restore the previous or a named generation \u2014 no approval, audited (CLI: local-ezai model rollback)",
        "tags": [
          "models"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "compose overlay (EZAI_CONTROL_PORT)",
      "url": "http://localhost:8010"
    }
  ]
}